split models.py in comps app into separate files
