+++ comps/management/commands/set_heat_level.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from comps.models import Heat
+from comps.models.heat import Heat
 
 class Command(BaseCommand):
     def handle(self, *args, **options):
