+++ comps/models/__init__.py
@@ -1,0 +1,5 @@
+from .comp import *
+from .heat_entry import *
+from .heat import *
+from .heatlist_dancer import *
+from .unmatched_heat_entry import *
