+++ comps/views/couple_heats.py
@@ -1,6 +1,7 @@
 from django.core.paginator import Paginator
 from django.shortcuts import render, get_object_or_404
-from comps.models import Comp, Heat
+from comps.models.comp import Comp
+from comps.models.heat import Heat
 from rankings.models import Couple
 from comps.filters import HeatFilter
 
@@ -9,7 +10,7 @@ def couple_heats(request, comp_id, couple_id):
     comp = get_object_or_404(Comp, pk=comp_id)
     couple = get_object_or_404(Couple, pk=couple_id)
 
-    heats = Heat.objects.filter(comp=comp).filter(heatentry__couple=couple).distinct().order_by('time')
+    heats = Heat.objects.filter(comp=comp).filter(heat_entry__couple=couple).distinct().order_by('time')
     paginator = Paginator(heats, 16)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
