+++ comps/views/couples.py
@@ -1,7 +1,8 @@
 from django.core.paginator import Paginator
 from django.db.models import Q
 from django.shortcuts import render, redirect, get_object_or_404
-from comps.models import Comp, HeatEntry
+from comps.models.comp import Comp
+from comps.models.heat_entry import Heat_Entry
 from rankings.models import Dancer, Couple
 from rankings.filters import DancerFilter
 
@@ -11,9 +12,7 @@ def couples(request, comp_id):
     if comp.process_state in [Comp.DANCERS_LOADED, Comp.DANCER_NAMES_FORMATTED, Comp.HEATS_LOADED]:
         return redirect ('comps:resolve_dancers', comp_id)
     else:
-        couples = Couple.objects.filter(heatentry__heat__comp=comp).distinct().order_by('dancer_1__name_last')
-        #f = DancerFilter(request.GET, queryset=dancers)
-        #paginator = Paginator(f.qs, 16)
+        couples = Couple.objects.filter(heat_entry__heat__comp=comp).distinct().order_by('dancer_1__name_last')
         paginator = Paginator(couples,16)
         page_number = request.GET.get('page')
         page_obj = paginator.get_page(page_number)
