+++ comps/views/create_comp.py
@@ -1,12 +1,12 @@
 from django.shortcuts import render, redirect
-from comps.models import Comp
+from comps.models.comp import Comp
 from comps.forms import CompForm
 
 
 def createcomp(request):
     if not request.user.is_superuser:
         return render(request, 'rankings/permission_denied.html')
-        
+
     if request.method == "GET":
         return render(request, 'comps/createcomp.html', {'form':CompForm()})
     else:
