+++ comps/views/heat.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from comps.models import Heat, HeatEntry
+from comps.models.heat import Heat
+from comps.models.heat_entry import Heat_Entry
 from comps.forms import HeatForm
 
 
@@ -8,7 +9,7 @@ def heat(request, heat_id):
     show_edit_button = request.user.is_superuser
 
     heat = get_object_or_404(Heat, pk=heat_id)
-    entries = HeatEntry.objects.filter(heat=heat)
+    entries = Heat_Entry.objects.filter(heat=heat)
     if entries.count() > 0:
         if entries.first().points is not None:
             entries = entries.order_by('-points')
