+++ comps/views/load_dancers.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
-from comps.models import Comp, HeatlistDancer, UnmatchedHeatEntry
+from comps.models.comp import Comp
+from comps.models.heatlist_dancer import Heatlist_Dancer
+from comps.models.unmatched_heat_entry import Unmatched_Heat_Entry
 from comps.heatlist.file_based_heatlist import FileBasedHeatlist
 from comps.heatlist.comp_mngr_heatlist import CompMngrHeatlist
 from comps.heatlist.comp_organizer_heatlist import CompOrgHeatlist
@@ -13,11 +15,11 @@ def load_dancers(request, comp_id):
     if len(comp_objects) == 1:
         comp=comp_objects[0]
 
-    if HeatlistDancer.objects.count() > 0:
-        HeatlistDancer.objects.all().delete()
+    if Heatlist_Dancer.objects.count() > 0:
+        heatlist_dancerDancer.objects.all().delete()
 
-    if UnmatchedHeatEntry.objects.count() > 0:
-        UnmatchedHeatEntry.objects.all().delete()
+    if Unmatched_Heat_Entry.objects.count() > 0:
+        Unmatched_Heat_Entry.objects.all().delete()
 
     if comp.heatsheet_file:
             heatlist = FileBasedHeatlist()
@@ -33,7 +35,7 @@ def load_dancers(request, comp_id):
         heatlist.open(comp.heatsheet_url)
 
     for d in heatlist.dancers:
-        in_database = HeatlistDancer.objects.filter(name = d.name)
+        in_database = Heatlist_Dancer.objects.filter(name = d.name)
         if in_database.count() == 0:
             d.save()
