+++ comps/views/load_scoresheets.py
@@ -1,12 +1,12 @@
 from django.core import serializers
 from django.shortcuts import render, get_object_or_404
-from comps.models import Comp
+from comps.models.comp import Comp
 from comps.tasks import process_scoresheet_task
 
 def load_scoresheets(request, comp_id):
     if not request.user.is_superuser:
         return render(request, 'rankings/permission_denied.html')
-        
+
     comp_objects = Comp.objects.filter(pk=comp_id)
     if len(comp_objects) == 1:
         comp=comp_objects[0]
